API_Tbot/requirements.txt
pyTelegramBotAPI==4.12.0
python_telegram_bot==20.3
# sympy==1.11.1
# matplotlib==3.6.2
# configparser==5.0.2
# platformdirs==2.6.0